requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==20.3.0
Automat==20.2.0
black==20.8b1
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
coverage==5.5
cryptography==3.4.7
cssselect==1.1.0
Django==3.2
django-filter==2.4.0
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.10.1
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==1.1.1
more-itertools==8.7.0
moto==2.0.5
mypy-extensions==0.4.3
packaging==20.9
parsel==1.6.0
pathspec==0.8.1
pluggy==0.13.1
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-env==0.6.2
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.5.0
regex==2021.4.4
requests==2.25.1
responses==0.13.2
s3transfer==0.3.7
Scrapy==2.5.0
scrapy-djangoitem==1.1.1
scrapy-s3pipeline==0.7.0
scrapy-sqspipeline==1.0.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
w3lib==1.22.0
Werkzeug==1.0.1
xmltodict==0.12.0
zipp==3.4.1
zope.interface==5.4.0
